pyproject.toml
[project]
name = "grazer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "<3.13,>=3.11"
dependencies = [
    "torch==2.5.1",
    "transformers==4.46.2",
    "torchvision==0.20.1",
    "timm==1.0.11",
    "sentence_transformers==3.2.1",
    "xgboost==2.1.2",
    "aiohttp==3.11.2",
    "runpod==1.7.7",
    "atproto==0.0.53",
    "python-dotenv~=1.0.0",
    "rq==2.0.0",
    "requests==2.32.3",
    "pillow==11.0.0",
    "inflection==0.5.1",
    "zstandard==0.23.0",
    "websocket-client==1.8.0",
    "openai==1.54.4",
    "nest_asyncio==1.6.0",
    "pyre2==0.3.6",
    "sentry-sdk==2.19.0",
    "timm==1.0.11",
    "ray[client]>=2.41.0",
    "ray==2.43.0",
    "cachetools==5.5.1",
    "aiorwlock>=1.5.0",
    "sentry-sdk",
    "numpy",
    "uvicorn==0.34.0",
    "pytest==8.3.5",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "pytest-asyncio==0.25.3",
    "grpcio==1.71.0",
]

[dependency-groups]
dev = [
    "cython>=3.0.12",
    "ninja>=1.11.1.3",
    "setuptools>=75.8.2",
    "flake8>=7.1.2",
    "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm.scripts]

# Lint all
lint = "flake8"

# Test all
test = "pytest"

# Example:
# pdm img_build ghcr.io/grazer:$(git rev-parse --short HEAD)
img_build = { shell = "docker build -t {args: grazer:local} ." }

# Startup commands

# Intialize Ray
ray_init = { shell = "ray stop && ray start --head --dashboard-host=0.0.0.0 --dashboard-port 8080" }

# NOTE: The following commands start a process in the background and dump stdout to a nohup.out file in the root directory
# Cache Worker
start_cache_worker = { shell = "nohup python run_ray_cache.py --name cache:main --num-cpus 2 --num-gpus 0 &" }

# Network Worker
start_network_worker = { shell = "nohup python run_ray_network_worker.py --name network:main --num-workers 5 --num-cpus 0.1 --num-gpus 0 &" }

# GPU Worker
start_gpu_worker = { shell = "nohup python run_ray_gpu_worker.py --name gpu --num-workers 3 --num-cpus 0.5 --num-gpus 0.2 &" }

# CPU Worker
start_cpu_worker = { shell = "nohup python run_ray_cpu_worker.py --name cpu:main --num-workers 6 --num-cpus 0.5 --num-gpus 0 &" }

# Runpod Worker
start_runpod_worker = { cmd = ["python3", "run_runpod_worker.py"] }

tail_output = { shell = "tail -f -n 300 nohup.out" }

# Starts all ray workers and tails the log output
start_all_workers.composite = [
    "ray_init",
    "start_cache_worker",
    "start_network_worker",
    "start_cpu_worker",
    "start_gpu_worker",
    "tail_output",
]

start_all_workers.env = { RAY_DEDUP_LOGS = "0", REDIS_URL = "redis://localhost:6379", SENTRY_DSN = "https://foo:bar@example.com/12345" }
start_workers = { shell = "ray job submit --runtime-env --working-dir $HOME/G/grazer -- python run_ray_cache.py", env = { RAY_DEDUP_LOGS = "0", REDIS_URL = "redis://localhost:6379", SENTRY_DSN = "https://foo:bar@example.com/12345", RAY_ADDRESS = "http://127.0.0.1:8265" } }


# NEW start workers in a Ray Cluster
test_workers = { cmd = "python start_cluster_workers.py", env = { RAY_DEDUP_LOGS = "0", REDIS_URL = "redis://localhost:6379", SENTRY_DSN = "https://foo:bar@example.com/12345" } }
